StartSystemForScattering in PotentialModule.f90
the subroutine has been redesigned in view of the recent
changes in the initialization of the trajectories
now it defines only asymptotic conditions for the
subsystem potential (which are used for minimum finding
in ScatteringSimulation) or starting conditions
for the scattering cordinate, which are used when
a new scattering trajectory is started.
In future developments the initialization subroutine
of ScatteringSimulation will be split so that
minimum and normal modes are computed only once at
the beginning of the calculation.